clean up main.go
